minor refactoring

+++ roxanne_rosjava_core/src/main/java/com/github/roxanne_rosjava/roxanne_rosjava_core/ai/deliberative/solver/PseudoControllabilityAwareSolver.java
@@ -15,6 +15,7 @@ import com.github.roxanne_rosjava.roxanne_rosjava_core.ai.framework.microkernel.
 import com.github.roxanne_rosjava.roxanne_rosjava_core.ai.framework.microkernel.lang.ex.PlanRefinementException;
 import com.github.roxanne_rosjava.roxanne_rosjava_core.ai.framework.microkernel.lang.flaw.Flaw;
 import com.github.roxanne_rosjava.roxanne_rosjava_core.ai.framework.microkernel.resolver.ex.UnsolvableFlawException;
+import com.github.roxanne_rosjava.roxanne_rosjava_core.ai.framework.utils.properties.FilePropertyReader;
 import org.bson.Document;
 
 import com.mongodb.MongoClient;
@@ -50,15 +51,27 @@ public class PseudoControllabilityAwareSolver extends Solver
 	public SearchSpaceNode solve() 
 			throws NoSolutionFoundException
 	{
-		// create a collection to the DB
-		MongoClient client = new MongoClient();
-		// get db 
-		MongoDatabase db = client.getDatabase("roxanne_search");
-		// get collection
-		MongoCollection<Document> collection = db.getCollection("search_data");
-		// remove all data from the collection
-		collection.drop();
-		
+		// set client connection
+		MongoClient client = null;
+		// prepare collection
+		MongoCollection<Document> collection = null;
+		// get deliberative property file
+		FilePropertyReader properties = new FilePropertyReader(
+				FRAMEWORK_HOME + FilePropertyReader.DEFAULT_DELIBERATIVE_PROPERTY);
+		// get mongodb
+		String mongodb = properties.getProperty("mongodb").trim();
+		// check if exists
+		if (mongodb != null && !mongodb.equals(""))
+		{
+			// create a collection to the DB
+			client = new MongoClient();
+			// get db
+			MongoDatabase db = client.getDatabase(mongodb);
+			// get collection
+			collection = db.getCollection("search_data");
+			// remove all data from the collection
+			collection.drop();
+		}
 		
 		// set solving start time
 		long start = System.currentTimeMillis();
@@ -108,27 +121,31 @@ public class PseudoControllabilityAwareSolver extends Solver
 				}
 				// info log 
 				info(info);
-				
-				
-				// create solving statistic record
-				Document doc = new Document("step", node.getId());
-				doc.append("fringe-size", this.fringe.getFringeSize());
-				doc.append("node-depth", node.getDepth());
-				doc.append("node-cost", node.getCost());
-				doc.append("node-heuristic", node.getHeuristic());
-				doc.append("node-min-makespan", node.getMakespan()[0]);
-				doc.append("node-max-makespan", node.getMakespan()[1]);
-				doc.append("node-min-duration", node.getBehaviorDuration()[0]);
-				doc.append("node-max-duration", node.getBehaviorDuration()[1]);
-				doc.append("node-number-of-flaws", node.getNumberOfFlaws());
-				// check additional metric
-				if (node.getDomainSpecificMetric() != null) {
-					// add additional metric
-					doc.append("node-domain-specific-metric", node.getDomainSpecificMetric().toString());
+
+
+				// check db collection
+				if (collection != null)
+				{
+					// create solving statistic record
+					Document doc = new Document("step", node.getId());
+					doc.append("fringe-size", this.fringe.getFringeSize());
+					doc.append("node-depth", node.getDepth());
+					doc.append("node-cost", node.getCost());
+					doc.append("node-heuristic", node.getHeuristic());
+					doc.append("node-min-makespan", node.getMakespan()[0]);
+					doc.append("node-max-makespan", node.getMakespan()[1]);
+					doc.append("node-min-duration", node.getBehaviorDuration()[0]);
+					doc.append("node-max-duration", node.getBehaviorDuration()[1]);
+					doc.append("node-number-of-flaws", node.getNumberOfFlaws());
+					// check additional metric
+					if (node.getDomainSpecificMetric() != null) {
+						// add additional metric
+						doc.append("node-domain-specific-metric", node.getDomainSpecificMetric().toString());
+					}
+
+					// insert data into the collection
+					collection.insertOne(doc);
 				}
-				
-				// insert data into the collection
-				collection.insertOne(doc);
 
 				
 				// propagate extracted node
